% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignComp.R
\name{assignComp}
\alias{assignComp}
\title{Assign pens to compartments}
\usage{
assignComp(data, beh, nPC = 8, etime = 10)
}
\arguments{
\item{data}{data.table}

\item{beh}{character vector}

\item{nPC}{numeric}

\item{etime}{numeric}
}
\value{
data.table
}
\description{
Assign pens to compartments
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6)
animalsInTrial <- assignTreatment(animalsInTrial,c("Wit","Groen","Rood","Geel"))
animalsInTrial <- assignComp(animalsInTrial,c("Wit","Groen","Rood","Geel"))
}
