% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotZoo.R
\name{boxplotZoo}
\alias{boxplotZoo}
\title{Boxplots of treatments over phases in feed trial}
\usage{
boxplotZoo(
  data,
  var,
  treatment = "behandeling1",
  group1 = "faseW",
  group2 = "ronde",
  group3 = NULL,
  lab = "Treatment",
  lab2 = "Round",
  pen = "hok",
  xlab,
  ylab,
  colours = brewer.pal(length(unique(data[, get(treatment)])), "Set3")
)
}
\arguments{
\item{data}{data.table}

\item{var}{character name of zootechnical parameter of interest}

\item{treatment}{character name of treatment variable (default= "behandeling1")}

\item{group1}{character name of grouping variable 1 (default= "faseW")}

\item{group2}{character name of grouping variable 2 (default= "ronde")}

\item{group3}{character name of grouping variable 1 (default= NULL)}

\item{lab}{character name to be used in legend for grouping variable 1 (default = "Treatment")}

\item{lab2}{character name to be used in legend for grouping variable 2 (default = "Round")}

\item{pen}{character name for pen variable (default= "hok")}

\item{xlab}{character label for x-axis}

\item{ylab}{character label for y-axis}

\item{colours}{character vector names for colours for each treatments (defaults to selection of n(treatment) colours from "Set3" of RColorBrewer palette)}
}
\value{
plot
}
\description{
Boxplots of treatments over phases in feed trial
}
\examples{
groeidataLong <- makeLong(groeidata,2)
boxplotZoo(data=groeidataLong,var="Weight",xlab="Treatment groups per phase",ylab= "Weight (kg)")
}
