% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotZoo.R
\name{check_outlier}
\alias{check_outlier}
\title{Find outliers in vector}
\usage{
check_outlier(v, coef = 1.5)
}
\arguments{
\item{v}{numeric vector}

\item{coef}{numeric coefficient  to multiply with IQR (Default = 1.5)}
}
\value{
boolean vector
}
\description{
Find outliers in vector
}
