% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{createWeightClass}
\alias{createWeightClass}
\title{Create weight classes}
\usage{
createWeightClass(data, nWC = 3)
}
\arguments{
\item{data}{data.table}

\item{nWC}{numeric}
}
\value{
data.table
}
\description{
Create weight classes
}
