% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResult.R
\name{exportResult}
\alias{exportResult}
\title{Export the results}
\usage{
exportResult(dataIn, dataOut, filename = "biggen", fileLoc = "H:\\\\Downloads")
}
\arguments{
\item{dataIn}{data.table}

\item{dataOut}{data.table}

\item{filename}{character filename (default =  "biggen")}

\item{fileLoc}{character file location (default = "~/Downloads")}
}
\value{
no return, only creates file in fileLoc folder
}
\description{
Export the results
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsOutTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[2]]
animalsInTrial <- assignPens(animalsInTrial,6)
animalsInTrial <- assignTreatment(animalsInTrial,c("Wit","Groen","Rood","Geel"))
animalsInTrial <- assignComp(animalsInTrial)
exportResult(animalsInTrial,animalsOutTrial)
}
