% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groeidata}
\alias{groeidata}
\title{Animal zootechnical data of ILVO trial}
\format{
A data frame with 192 rows and 10 variables:
\describe{
\item{bedrijf}{Farm}
\item{ronde}{Round}
\item{compartiment}{Compartment}
\item{hok}{Pen}
\item{behandeling1}{Treatment 1}
\item{behandeling2}{Treatment 2}
\item{sex}{Sex}
\item{oornr}{ID}
\item{geboorte_di}{Birth date}
\item{geboorte_gi}{Birth weight}
\item{moeder}{Mother ID}
\item{beer}{Beer}
\item{speen_d}{Weaning date}
\item{speen_gi}{Weaning weight}
\item{w1_d}{Date first weighing}
\item{w1_gi}{Weight at first weighing}
\item{w1_vo}{First feed weight}
\item{w1_aantal}{Number of piglets at first weighing}
\item{w2_d}{Date second weighing}
\item{w2_gi}{Weight at second weighing}
\item{w2_vo}{Second feed weight}
\item{w2_aantal}{Number of piglets at second weighing}
\item{igroei1}{Growth in first phase}
\item{igroei2}{Growth in second phase}
\item{igroeitot}{Total growth}
\item{g1}{Weight after first phase}
\item{dg1}{Daily gain during first phase}
\item{dvo1}{Daily feed intake during first phase}
\item{vc1}{Feed conversion during first phase}
\item{g2}{Weight after second phase}
\item{dg2}{Daily gain during second phase}
\item{dvo2}{Daily feed intake during second phase}
\item{vc2}{Feed conversion during second phase}
\item{gbegin}{Weight at start}
\item{geinde}{Weight at end}
\item{dgtot}{Daily gain during total trial}
\item{dvotot}{Daily feed intake during total trial}
\item{vctot}{Feed conversion during total trial}
\item{cvbegin}{Coefficient of variance at start}
\item{cv1}{Coefficient of variance after first phase}
\item{cv2}{Coefficient of variance after second phase}
}
}
\source{
ILVO trial data
}
\usage{
groeidata
}
\description{
A dataset containing the weights, feed intake and derived parameters of piglets in an ILVO feed trial.
}
\keyword{datasets}
