% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLong.R
\name{makeLong}
\alias{makeLong}
\title{Put zootechnical data in suitable format}
\usage{
makeLong(
  data,
  n_fasenW,
  vars = c("g", "dg", "dvo", "vc"),
  VARfasenames = c("Weight", "Daily_gain", "Daily_feed_intake",
    "Feed_conversion_ratio")
)
}
\arguments{
\item{data}{data.table of zootechnical data}

\item{n_fasenW}{numeric number of phases in the trial}

\item{vars}{character vector of variable names for zootechnical parameters in data}

\item{VARfasenames}{character vector of names to be used for these parameters in reshaped data.frame}
}
\value{
data.table with zootechnical parameters in long format over phases
}
\description{
Put zootechnical data in suitable format
}
\examples{
groeidataLong <- makeLong(groeidata,2)
}
