% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsZoo.R
\name{modelsZoo}
\alias{modelsZoo}
\title{Title}
\usage{
modelsZoo(
  data,
  precision = 2,
  VARnames,
  effects,
  treatment = "behandeling1",
  group1 = NULL
)
}
\arguments{
\item{group1}{}
}
\value{

}
\description{
Title
}
