% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{nb}
\alias{nb}
\title{Select siblings to switch}
\usage{
nb(x, data, sowids)
}
\arguments{
\item{x}{numeric vector of pens}

\item{data}{data.table with column Zeugnr (sow ID)}

\item{sowids}{character vector of sow ID's that appear more than once (i.e. siblings)}
}
\value{
numeric vector
}
\description{
Selects and switches siblings randomly
}
