% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{OF}
\alias{OF}
\title{Function to optimize for switchInPen function}
\usage{
OF(x, data, ...)
}
\arguments{
\item{x}{numeric pen number}

\item{data}{data.table}

\item{...}{parameters to be applied to other functions}
}
\value{
numeric
}
\description{
Function to optimize for switchInPen function
}
