% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTrialAnimals.R
\name{selectTrialAnimals}
\alias{selectTrialAnimals}
\title{Select trial animals}
\usage{
selectTrialAnimals(data, nB, nZ, minWeight, maxWeight)
}
\arguments{
\item{data}{data.table}

\item{nB}{numeric}

\item{nZ}{numeric}

\item{minWeight}{numeric}

\item{maxWeight}{numeric}
}
\value{
list
}
\description{
Select which animals to include and which to exclude from the trial
}
\examples{
selectTrialAnimals(biggen,72,72,5.4,9.5)
}
