% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{switchInPen}
\alias{switchInPen}
\title{Avoid siblings in pens}
\usage{
switchInPen(data)
}
\arguments{
\item{data}{data.table}
}
\value{
data.table
}
\description{
Switches siblings in same pen to other pen while keeping standard deviation in average weight per pen to a minimum
}
