% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResult.R
\name{trialSummary}
\alias{trialSummary}
\title{Give summary statistics}
\usage{
trialSummary(data)
}
\arguments{
\item{data}{data.table}
}
\value{
no return, only prints summary
}
\description{
Give summary statistics
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6)
animalsInTrial <- assignTreatment(animalsInTrial,c("Wit","Groen","Rood","Geel"))
animalsInTrial <- assignComp(animalsInTrial)
trialSummary(animalsInTrial)
}
