% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignComp.R
\name{assignComp}
\alias{assignComp}
\title{Assign pens to compartments}
\usage{
assignComp(data, nPC = 8, etime = 10)
}
\arguments{
\item{data}{data.table with columns Sex (sex of piglets), Gew_klasse (weight class), Beh (Treatment), and Hok (Pen)}

\item{nPC}{numeric number of pens per compartment}

\item{etime}{numeric number of seconds to allow iterations before dropping the stringency for level of balance to strive for in weight class distribution over compartments}
}
\value{
data.table with added column Comp (compartment)
}
\description{
Assign pens to their compartment randomly but balanced over sex, treatment, and as much as possible over weight class
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6)
animalsInTrial <- assignTreatment(animalsInTrial,c("Wit","Groen","Rood","Geel"))
animalsInTrial <- assignComp(animalsInTrial)
}
