% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{assignPens}
\alias{assignPens}
\title{Assign piglets to pens}
\usage{
assignPens(data, nH, cutoff, nWC = 3)
}
\arguments{
\item{data}{data.table}

\item{nH}{numeric}

\item{cutoff}{numeric}

\item{nWC}{numeric}
}
\value{
data.table
}
\description{
Assign piglets to pens
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6,0.1)
}
