% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{assignPens}
\alias{assignPens}
\title{Assign piglets to pens}
\usage{
assignPens(data, nH, nWC = 3, separateSex = T)
}
\arguments{
\item{data}{data.table with columns Zeugnr (sow ID),  Sex (sex of piglets), and Speen_gew (weaning weights of piglets)}

\item{nH}{numeric number of piglets to be housed in each pen}

\item{nWC}{numeric number of weight classes to be created (default = 3)}

\item{separateSex}{boolean Should barrows and gilts be kept in separate pens (default = TRUE)? If FALSE, the number of barrows and gilts should be equal in total and per pen.}
}
\value{
data.table with added column Hok (pen)
}
\description{
Assign piglets to their pens keeping standard deviation in average weight per pen to a minimum.
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6)
}
