% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTreatment.R
\name{assignTreatment}
\alias{assignTreatment}
\title{Assign treatments to pens}
\usage{
assignTreatment(data, beh)
}
\arguments{
\item{data}{data.table}

\item{beh}{character vector}
}
\value{
data.table
}
\description{
Assign treatments to pens
}
\examples{
animalsInTrial <- selectTrialAnimals(biggen,72,72,5.4,9.5)[[1]]
animalsInTrial <- assignPens(animalsInTrial,6,0.1)
animalsInTrial <- assignTreatment(animalsInTrial,c("Wit","Groen","Rood","Geel"))
}
