% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{createWeightClass}
\alias{createWeightClass}
\title{Create weight classes}
\usage{
createWeightClass(data, nWC = 3)
}
\arguments{
\item{data}{data.table}

\item{nWC}{numeric number of weight classes to assign can be 1, 2, 3 or 4}
}
\value{
data.table
}
\description{
Assigns weight class to each piglet
}
