% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsZoo.R
\name{modelsZoo}
\alias{modelsZoo}
\title{Run models for each parameter}
\usage{
modelsZoo(
  data,
  precision = 2,
  VARnames,
  effects,
  treatment = "behandeling1",
  group1 = NULL
)
}
\arguments{
\item{data}{data.table}

\item{precision}{numeric number of digits after the decimal}

\item{VARnames}{character vector of variables to use from data}

\item{effects}{formula for starting model}

\item{treatment}{character name of treatment variable}

\item{group1}{character name of second grouping variable, typically faseW}
}
\value{
table of means and sd with significance codes
}
\description{
Run models for each parameter
}
\examples{
groeidataLong <- makeLong(groeidata,2)
tabel <- modelsZoo(groeidataLong,VARnames = c("geinde","dgtot"),effects = "behandeling1+gbegin+(1|compartiment/ronde)")
}
