% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{nb}
\alias{nb}
\title{Select siblings to switch}
\usage{
nb(x, data, sowids)
}
\arguments{
\item{x}{numeric vector of pens}

\item{data}{data.table}

\item{sowids}{character vector}
}
\value{
numeric vector
}
\description{
Select siblings to switch
}
