% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPens.R
\name{switchInPen}
\alias{switchInPen}
\title{Avoid siblings in pens}
\usage{
switchInPen(data, cutoff)
}
\arguments{
\item{data}{data.table}

\item{cutoff}{numeric}
}
\value{
data.table
}
\description{
Avoid siblings in pens
}
